moved getTokens to useHome, added try catch to methods
